// lib/types.ts
export type WorkType = "short" | "medium" | "feature";
export type WorkSection =
  | "competition_int"
  | "competition_nat"
  | "short_competition"
  | "panorama"
  | "special"
  | "retrospective"
  | "tribute";
// ...
export interface ScreeningWork {
  id: number;
  title: string;
  type: WorkType;
  section: WorkSection;
  director: string;
  runtime: number;
  still: string;
}

export interface Screening {
  id: number;
  edition: number;
  work: ScreeningWork;
  work_id: number;
  venue: number | null;
  venue_name: string | null;
  date: string;
  time: string;
  notes: string;
}